+++ src/bacnet/basic/object/color_object.c
@@ -704,7 +704,9 @@ bool Color_Object_Name(
             status =
                 characterstring_init_ansi(object_name, pObject->Object_Name);
         } else {
-            snprintf(name_text, sizeof(name_text), "COLOR-%u", object_instance);
+            snprintf(
+                name_text, sizeof(name_text), "COLOR-%lu",
+                (unsigned long)object_instance);
             status = characterstring_init_ansi(object_name, name_text);
         }
     }
