+++ src/bacnet/basic/object/color_temperature.c
@@ -944,8 +944,8 @@ bool Color_Temperature_Object_Name(
                 characterstring_init_ansi(object_name, pObject->Object_Name);
         } else {
             snprintf(
-                name_text, sizeof(name_text), "COLOR-TEMPERATURE-%u",
-                object_instance);
+                name_text, sizeof(name_text), "COLOR-TEMPERATURE-%lu",
+                (unsigned long)object_instance);
             status = characterstring_init_ansi(object_name, name_text);
         }
     }
