+++ src/bacnet/basic/object/lo.c
@@ -983,8 +983,8 @@ bool Lighting_Output_Object_Name(
                 characterstring_init_ansi(object_name, pObject->Object_Name);
         } else {
             snprintf(
-                name_text, sizeof(name_text), "LIGHTING-OUTPUT-%u",
-                object_instance);
+                name_text, sizeof(name_text), "LIGHTING-OUTPUT-%lu",
+                (unsigned long)object_instance);
             status = characterstring_init_ansi(object_name, name_text);
         }
     }
